wip: usb: Fix ioctl calls
It looks like the user-space passes more data then we actually expect
from ioctl definition.

So we use a larger buffer to copy the required amount of data and cast
the used buffer to the expected structure. To make the buffer handling
easier we perform the initial copy_form_user before we enter the switch.

We also have to replace some memcpy calls inside the specialized ioctl
sub-calls with copy_from_user.

Signed-off-by: Soeren Grunewald <soeren.grunewald@desy.de>